+++ tests/integ/test_kmeans.py
@@ -21,7 +21,7 @@ import time
 import pytest
 
 from sagemaker import KMeans, KMeansModel
-from sagemaker.utils import name_from_base
+from sagemaker.utils import unique_name_from_base
 from tests.integ import DATA_DIR, TRAINING_DEFAULT_TIMEOUT_MINUTES
 from tests.integ.timeout import timeout, timeout_and_delete_endpoint_by_name
 
@@ -64,7 +64,7 @@ def test_kmeans(sagemaker_session):
 
         kmeans.fit(kmeans.record_set(train_set[0][:100]))
 
-    endpoint_name = name_from_base('kmeans')
+    endpoint_name = unique_name_from_base('kmeans')
     with timeout_and_delete_endpoint_by_name(endpoint_name, sagemaker_session):
         model = KMeansModel(kmeans.model_data, role='SageMakerRole', sagemaker_session=sagemaker_session)
         predictor = model.deploy(1, 'ml.c4.xlarge', endpoint_name=endpoint_name)
@@ -78,7 +78,7 @@ def test_kmeans(sagemaker_session):
 
 def test_async_kmeans(sagemaker_session):
     training_job_name = ""
-    endpoint_name = name_from_base('kmeans')
+    endpoint_name = unique_name_from_base('kmeans')
 
     with timeout(minutes=5):
         data_path = os.path.join(DATA_DIR, 'one_p_mnist', 'mnist.pkl.gz')