+++ tests/integ/test_ntm.py
@@ -19,7 +19,7 @@ import pytest
 
 from sagemaker import NTM, NTMModel
 from sagemaker.amazon.common import read_records
-from sagemaker.utils import name_from_base
+from sagemaker.utils import unique_name_from_base
 from tests.integ import DATA_DIR, TRAINING_DEFAULT_TIMEOUT_MINUTES
 from tests.integ.timeout import timeout, timeout_and_delete_endpoint_by_name
 from tests.integ.record_set import prepare_record_set_from_local_files
@@ -44,7 +44,7 @@ def test_ntm(sagemaker_session):
                                                          len(all_records), feature_num, sagemaker_session)
         ntm.fit(record_set, None)
 
-    endpoint_name = name_from_base('ntm')
+    endpoint_name = unique_name_from_base('ntm')
     with timeout_and_delete_endpoint_by_name(endpoint_name, sagemaker_session):
         model = NTMModel(ntm.model_data, role='SageMakerRole', sagemaker_session=sagemaker_session)
         predictor = model.deploy(1, 'ml.c4.xlarge', endpoint_name=endpoint_name)